+++ package.json
@@ -27,7 +27,7 @@
     "expo-haptics": "^14.1.4",
     "expo-image-picker": "^16.1.4",
     "expo-linking": "^7.1.5",
-    "expo-router": "~5.1.5",
+    "expo-router": "5.1.5",
     "expo-splash-screen": "^0.30.8",
     "expo-status-bar": "^2.2.3",
     "expo-symbols": "^0.4.4",
@@ -44,7 +44,7 @@
     "react-native-paper": "^5.13.1",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.11.1",
+    "react-native-screens": "4.11.1",
     "react-native-web": "^0.20.0",
     "react-native-webview": "13.13.5"
   },
